Move tests to test_c10d_spawn
